Make some big improvements to the code

+++ engine/nine_morris_3d_engine/audio/sound_data.h
@@ -18,7 +18,7 @@ public:
     int get_samples() { return samples; }
     int get_channels() { return channels; }
     int get_sample_rate() { return sample_rate; }
-    int get_bits_per_sample() { return bits_per_sample; }
+    size_t get_bits_per_sample() { return bits_per_sample; }
 	std::string_view get_file_path() { return file_path; }
 private:
     size_t compute_size();
